Remove jekyll
This is a complete rewrite of my blog to a vastly simpler form. That is,
we only need HTML and a couple of build scripts to have something
equivalent, without all the dependencies that jekyll brings. I have also
taken the chance to remove a lot of stuff that was actually not needed.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>